libtomlc99: add simple TAP test
The tomlc99 project includes a copy of the general TOML
test suite: https://github.com/BurntSushi/toml-test

In order to avoid taking on its Golang dependency,
we simply add a small TAP test that runs all the valid and
invalid TOML input files from the BurntSushi suite through
the libtomlc99 parser and ensure that they pass/fail as expected.

A few invalid input files are not handled properly.  Add a
table of them with the reason for failure so the TAP test
can skip them and emit the reason for skipping.