lib/sigcert: add signing certificate class
RFC 15 calls for users to sign J so that other users,
including the instance owner, cannot forge run requests
to the IMP.

One mechanism that satisfies this requirement is
libsodium's public key signatures based on ed25519.
This is a good starting point since we are already using
an instance owner's CURVE keys in Flux to protect
overlay network peer connections, and thus are already
dependent on libsodium.

0MQ introduced a certificate format for its socket
security called zcert (czmq.zeromq.org/manual:zcert).
These certificates cannot be used for public key
signatures because libsodium uses a different key format
for signing keys.  Thus, we introduce a new class
called flux_sigcert that is similar to zcert, but not
dependent on czmq.  The on-disk certificate format
uses TOML instead of ZPL format, and base64 encoding
instead of Z85 for ASCII armor.

This class is intended to be expanded to support JSON
signing, certificate signing, and certificate serialization
as JSON, however in this iteration only detached blob signing
is supported.