lib/sigcert: avoid use of uninitialized secret key
Problem: signing without access to secret key succeeds,
deferring error until verification.

This idiom was copied from CZMQ "zcert" class, and I think
the intent was to keep the class as simple as possible.
The downside is that errors are not detected until signature
verification, possibly when the signed material has already
been transferred to another program, where it may be difficult
to inform the user of the precise error.

This would seem to impact usability for us.  For example, a
user submitting an array of jobs may load the certificate
once, sign many job requests, and not find out until the jobs
are ingested that the signatures are invalid.

Change the code so that:
1) flux_sigcert_store() fails if secret key is invalid
2) flux_sigcert_equal() succeeds if both secret keys are invalid
3) flux_sigcert_sign() fails if secret key is invalid