lib/sigcert: add metadata to cert
Problem: users of the cert class will need to store and
retrieve auxiliary data in the certificate, but we have
no way to do that.  For example, a CA may need to include
an expiration date in the cert, or a keygen utility
may want to add its version.

Add a simple hash of string values to the cert object, and
add get/set methods to its public interface.  When serialized
to TOML/JSON, metadata values appear in their own [metadata]
stanza/object.

This idea was borrowed from the CZMQ zcert class.

Add unit tests.