lib/sigcert: make meta functions type specific
Problem: only strings can be represented as sigcert metadata.

Rename
  flux_sigcert_meta_get() -> flux_sigcert_meta_gets()
  flux_sigcert_meta_set() -> flux_sigcert_meta_sets()

This will allow methods for other types to be added.

Change the prototype of "gets" to return the meta value in
a result argument rather than the function return value, so that
a similarly constructed function returning an integer can have a
way of returning an error.