lib/sigcert: support timestamp metadata
Add methods to support time_t-typed metadata:

flux_sigcert_meta_setts()
flux_sigcert_meta_getts()

The time is expected to be in UTC seconds-since-epoch
(as is returned from time(2).  It is serialized to
JSON as an object that looks like this
  { "iso-8601-ts" : "2003-08-24T05:14:50Z" }

It is converted to native TOML timestamps, which are
also in ISO 8601 format.

Add unit tests.