lib/sigcert: add certificate signature capability
Add functions for sign/verify of a cert with another cert:

flux_sigcert_sign_cert()
flux_sigcert_verify_cert()

The signature is generated by first serializing the target
cert, including metadata, and excluding secret key and
existing signature (if any), in a repeatable way.
This is signed with the private key of the signing cert
and stored in the target cert.

The cert can be serialized with its signature, and when
it is reconstituted, it can still be verified due to the
repeatability of the serialization.

This can be used to implement a certificate authority:
- CA authenticates user
- CA stores [metadata] like userid, cert id, valid date range
- CA signs cert (including metadata)
- (cert is transferred through untrusted path)
- cert can be verified with CA public key

Add unit tests.