lib/sigcert: call sodium_init() only once
Problem: sodium_init() is intended to be called once per
use of libsodium, not repeatedly, but we track whether it
has been called using a flag in each cert object.

Change that flag to a global variable so we only call it once.

It's apparently not harmful to call sodium_init() more than once,
and it is said by its docs to be thread-safe, so it's probably
not harmful to allow this flag ot be a bit "racey" with respect to
multiple threads calling our library.