lib/sigcert: define base64 buffers statically
Problem: there is are unnecessary mallocs and associated
error handling when keys are base64 encoded for serialization.

Add some fixed length typedefs for the secret keys, public
keys, and signatures so that we can use those on the stack
and in the cert structure.

Add simple encode/decode functions for those specific types
that should be hard to misuse.