lib/sigcert: only store secret-key in secret file
Problem: "cert" and "cert.pub" contain duplicate information
making it difficult to only update one of them as might
be needed when getting a public key signed.

Instead of having "cert" contain a superset of "cert.pub",
have it only contain secret-key.

Change flux_sigcert_load() so that "cert.pub" is loaded
unconditionally, and "cert" is only loaded if requested
by a new parameter.

Change flux_sigcert_store() so that "cert.pub" is stored
unconditionally, and "cert" is only stored if the secret
key in the object is valid.

Update tests.