libutil/cf: add config class
Add config class for TOML config file format.
This is a thin layer on top of a Jansson object, but
doesn't directly expose Jansson interfaces.

It supports buildig a config object from multiple TOML
files/buffers, checking it against a list of required/optional
keys and types, and accessing values.

It supports a hierarchy of tables so that config sections
for different components can coexist in a TOML file and
be broken down and checked/accessed separately.

After a table has been checked, values may be accessed
without further error checking.  If a key is misspelled in
such an access, a valid value is still returned (such as zero
or an empty string).