libutil/kv: serialize to "key\0val\0"
Problem: encoding of "key=val\0" complicates requires key
to be copied in order to return it to the user with null
termination when kv_entry_key() is called.

Encode as "key\0val\0" so key is NULL-terminated in situ.

This also eliminates the arbitrary limit on key length,
and requirement that key cannot contain the '=' character.