libutil/cf: add cf_update_glob()
Add function to call cf_update_file() on a cf table with a series of
files matching a globbing pattern. The current behavior of the function
is to only update the cf table if all files are successfully parsed.

This function is meant to be used to read in configuration split among
many files in a `conf.d/` style directory, e.g.

 cf_update_glob (cf, "/etc/conf.d/*.toml", NULL);

The number of files successfully parsed is returned (a return value of
zero means there were no matches, which may or may not be an error
condition).