libutil/test/kv: fix segfault
Problem: kv unit test prints bad data and occasionally segfaults.

The test in question essentially does this
  ok (foo (&s) == 0, "foo works and returned %s", s);

The libtap ok() macro passes the result of the test expression,
a format string, and variadic arguments down to a C function
ok_at_loc().  Since the order of function parameter evaulation
is not defined in C, one cannot assume that the test expression
is evaluated before the format and variadic arguments.

Drop the result output, which was extraneous to the purpose of
the test, from the ok() call.

Fixes #31.