libutil/sigcert: add attached sign/verify
Add support for attached signatures on strings.

sigcert_sign() appends a base64 signature to a string.
The signature is calculated over the initial contents
of the string, which must be NULL terminated.  A period
(".") is appended to the original string, followed by
the signature converted to base64.  The result is always
NULL terminated.  No memory is allocated - the signature
must fit in the specified buffer size or the call fails.

sigcert_sign_length() returns the amount of storage that
will be needed by sigcert_sign() to append the signature.

sigcert_verify() verifies an attached base64 signature
added by sigcert_sign() and returns the length of the
original string, without the appended period and signature
string.

This is intended to support a 'signer' class that encodes
signed payloads in a manner resembling JSC Compact Serializion,
described in Internet RFC 7515.