libutil/ca: add ca class
Add a class that can be used to construct certificate
authority tools.  It can be used to sign user certificates,
validate them, and revoke them.

A utility for signing user certs requires access to the CA
secret key.  Presumably it would be contained in a secure
environment and implement local controls to minimize the
chance of accidental disclosure of the CA secret key.
Such a utility would call:
  ca_create - ingest the [ca] configuration
  ca_load - (secret=true) load the CA cert from configured location
  ca_sign - set required metadata including userid to user cert,
            then sign it, accessing the CA secret key
  ca_sign - sign another one
  ...
  ca_destroy - teardown

Cert authentication requires access to the CA public key,
and may be embedded in unprivileged code, which would call:
  ca_create - ingest the [ca] configuration
  ca_load - (secret=false) load the CA cert from configured location
  ca_verify - verify cert was signed by CA public key
              and get userid
  ca_verify - verify another one
  ...
  ca_destroy - teardown

Cert revocation consists of adding a cert's UUID to a directory
that is propagated, probably from the protected cert signing
environment, to all environments where certs are authenticated.
Propagation is outside the scope of the class.