imp: add internal IMP state structure
Define an imp_state structure to hold internal IMP data. The struct
is defined in imp_state.h for use by other components.

For now the IMP state contains only the cmdline of the IMP at startup,
but in the future may be used for caching credentials, privilege
separation state, config file, etc.