imp: build: prepare for IMP config file
Link IMP with libtomlc99 to prepare for loading config file(s).

Create a built sources `testconfig.c` which contains the pattern
for the IMP configuration for the build directory version of the IMP.
Later, a separate "installed" flux-imp can be built as well with
a different built-in configuration pattern which points to the
system config paths.