sharness: run privsep unit test under sudo
The privsep unit test must be run under sudo. Add a sharness test
that invokes this unit test (and any other future unit tests that
may require sudo) when the SUDO prereq is met. This is not ideal
since the TAP output of the test(s) will be suppressed, but it is
slightly better than not running tests at all.