imp: add privilege separation and subcmd support
Initialize privilege separation in the IMP when running setuid.

The unprivileged child created during privilege separation then
tries to pull a single subcommand from the IMP's command line and
runs the unprivilged portion of that subcommand.

Currently the privileged parent does nothing but wait for child
to exit by virtue of closing its privsep pipe.

The only subcommand supported for now is the "version" command,
which is already registered in `impcmd.c`.