imp: link testconfig.c with the test IMP
Generate the default built-in config pattern in testconfig.h instead of
testconfig.c, as testconfig.c now contains a function to return the value
instead of just exporting the hard-coded global.

Link the IMP with testconfig.c so that both the global and the new function
returning the config pattern are available.