lib/sign_none: add "none" mechanism
The "none" mechanism is intended for testing as well
as possible use in a single-user instance.

The signature is always "none".  Verification consists
of checking that, and checking that the header userid
is the same as the real userid.

The userid check allows "none" to be used to sign
requests in a single user instance.  If other measures
prevent users other than the instance owner from submitting
jobs, then the instance itself has no need to securely
validate a job request.  The IMP, running setuid root
but with the real userid of the invoking user, can accept
a userid claim that matches the real userid since it is
not being asked to run as a user other than the one that
invoked it.