configure: enable maintainer mode by default
Add "enable" option to AM_MAINTAINER_MODE macro. This enables
maintainer-mode by default, but still allows for the use of
--disable-maintainer-mode by packagers and users when necessary.
(I think this is required for debian packaging for example)

Use of AM_MAINTAINER_MODE is now discouraged, and if used at all
adding the "enable" option is strongly encouraged, so that the
package will build as consistently as possibly by default.

See:
https://www.gnu.org/software/automake/manual/html_node/maintainer_002dmode.html
https://blogs.gnome.org/desrt/2011/09/08/am_maintainer_mode-is-not-cool/
https://autotools.io/automake/maintainer.html