autogen.sh: do not move aclocal.m4 out of top dir
Problem: The autogen.sh script copies aclocal.m4 into config/,
assuming it will be found there by Makefiles. However, this is
not the case, and when maintainer-mode is enabled, the "missing"
aclocal.m4 causes a rebuild and re-run of configure.

I'm not sure how to convince automake that aclocal.m4 can be found
under config, and since we have to copy it by hand that implies
we are probably doing something wrong, so just leave it alone.