testsuite: fix t1000-imp-basic.t tests on nosuid filesystems
Problem: The setuid-based tests in t1000-imp-basic.t fail when
the user has sudo access, but is running the testsuite on a filesystem
with nosuid, or where setuid permissions are otherwise not allowed.
In this case the `chmod` succeeds, but the permissions will be
silently unchanged.

Detect this condition by ensuring the resulting flux-imp copy has
exactly setuid perms (4755). If so, then set a SUID_ENABLED prereq
used by the follow-on tests.