imp: add minimal flux-imp exec subcommand
Add a minimally functional implementation of flux-imp exec.

This version takes JSON input on stdin with a single required key, J.
If J has a valid signature, then the IMP transitions to the signing
userid and executes a program with single argument given on the
command line.

In order to work, a new "[exec]" table should be added to the IMP
configuration. This table should have at a minimum the following
keys:

 - "allowed-users":  a list of usernames allowed to run flux-imp exec.
 - "allowed-shells": a list of allowed shell paths the IMP will run.

Still todo:
 - clean environment before executing shell as user.