imp: fix potential fd leak in privsep_destroy
Problem: privsep_destroy checks for fds greater than zero before
closing the parent and child pipe fds, but 0 is a valid file
descriptor.  This could lead to an fd leak.

Change the test in privsep destroy to check for fds >= 0 before
calling close to close the leak.

Additionally, initialize the wfd and rfd members of the privsep object
to -1 instead of 0 so privsep_destroy() doesn't accidentally close
stdin file descriptor.