lib: add flux_sign_wrap_as()
Add a new variant of flux_sign_wrap(3) to libflux-security which
allows the encoded userid to be passed into the function, instead
of obtained via getuid(). The old flux_sign_wrap(3) now just calls
flux_sign_wrap_as (ctx, getuid(), ...).

This function may be useful for the implementation of tests in
flux-framework projects that use libflux-security.