imp: disable exec in unprivileged mode by default
Problem: Allowing the IMP to fall back to executing its arguments
in unprivileged mode may pose a slight security risk. For example,
if the IMP is installed along with a system instance of Flux running
as user 'flux', but the IMP is not installed setuid, then any user
can run anything as user 'flux' by default.

Prevent this scenario by requiring that unprivileged IMP exec be
explicitly configured with a new exec parameter:

 allow-unprivileged-exec

which is false by default.