imp: add a kill command to the IMP
The IMP kill command allows an unprivileged user to send signals
to muiltiuser jobs which were started by an instance of the same
user.

This version of flux-imp kill assumes a Flux system instance was
started under systemd with Delegate=yes. This means all multi-user
will be running under a cgroup owned by the Flux instance owner.

The kill command will then authorize a user to kill any process that
is running within a cgroup owned by the requesting userid.