imp: kill: allow pid owner to kill their own process
For flux-imp kill, also fetch the PID owner by running stat(2) on
the /proc/<pid> directory. If the current real userid matches
the owner of the process, allow the IMP to send the signal.

This is mostly useful in testing flux-imp kill when not running
under flux, so cgroup ownership of all processes is likely root.