libutil: add kv_expand_environ/kv_environ_destroy
Problem: It would be convenient if a `struct kv` could be used to
store a set of environment variables, but this is not supported
by this interface.

Add kv_expand_environ() which returns a new  `char **environ` array
from a `struct kv` object.

The returned environment array muts be freed by kv_environ_destroy().