libutil: add strlcpy
Problem: strncpy(3) is a frequent source of programmer mistakes
which directly result in security vulnerabilities. In order to be
used safely, strncpy(3) requires an initialized buffer, or manual NUL
termination, plus manual checks for length to ensure a result buffer
without NUL termination is not created. This should not have to be
the case in a project where the purpose is to create secure code.

strlcpy(3), on the other hand, is designed "to be a safer, more
consistent, and less error prone replacement" for strncpy(3) [1]. With
its guarantee of NUL termination and safer semantics, it behooves us
to include this implementation in the flux-security project.

Note that since strlcpy(3) is known not be included in glibc, we can
include it wholesale here. However, when porting to BSD systems
presence of strlcpy() in libc may need to be detected and this
source file excluded from the build if an implementation is found.

[1] https://man.openbsd.org/strlcpy