lib/test: update invalid flag test in test_context.t
Problem: The test for invalid flag set during flux_security_create()
uses "1" as the invalid flag, but this will be the first _valid_ flag
added when flags are supported in this call, so the test will break
immediately.

Instead use 128 (0x80), which should be unlikely to be a valid flag
in the forseeable future.