configure: set CODE_COVERAGE_ENABLED in config.h
Problem: There is no current way to tell if flux-security was built
with --enable-code-coverage so that coverage specific actions can
be taken, e.g. calling __gcov_flush().

If coverage is enabled, set CODE_COVERAGE_ENABLED in config.h
so that code can use it to run conditional code under the coverage
build.