libutil: kv: add helpers for encode of argument vectors
Problem: flux-security utilities like the flux-imp may want to encode
and transmit argument vectors via a `struct kv`, but there is currently
no easy mechanism to enable this use case.

Add some functions to libutil/kv which allow the easy encoding and
decoding of argument vectors to and from a struct kv.