imp: exec: allow more than one argument to job shell
Problem: The IMP exec functionality currently only allows a single
argument to the job shell, presumed to be the target jobid. This
limits the enclosing instance from configuring other, default job
shell options -- for example to allow an alternate default initrc
location or a reconnect option.

Encode all arguments passed to flux-imp in the 'struct kv' passed to
the privileged parent, and use this argv in execvp(2). This allows
an arbitrary number of extra options to passed along to the launched
job shell.