testsuite: fix spellcheck
Problem: 'make check' does not fail on misspelled words.

Aspell doesn't like dictionary words that end in numbers.
The spellcheck script successfully runs an aspell "dry run" but
does so without the dictionary.  Later, when it checks for misspellings,
it is counting lines of output and ignores the return value of aspell.

Add the dictionary to the dry run so it fails there if there are problems.
Remove IPv4 and IPv6 from the dictionary.
Fix misc. spelling errors that are now caught and add some words to dictionary.