imp: kill: support systems where systemd uses cgroups v2
Problem: The flux-imp kill command does not support systemd running
with unified cgroups (cgroups v2), since the code assumes the cgroup
filesystem will be mounted at /sys/fs/cgroup/systemd (a legacy
mount point).

Based on information in https://systemd.io/CGROUP_DELEGATION/ and
systemd code itself, add a function to determine if a unified cgroup
hierarchy is mounted and its mount point, and use this information
in determining if the targeted PID is allowed to be signaled by the
requesting user.

This should allow the IMP kill command to work on most, if not all,
extant systemd installations.

Fixes #141