imp: exec: initial linger support
Problem: The implementation of `flux-imp exec` currently directly
executes the target job shell, replacing the IMP process entirely.
This was a purposeful decision to limit the amount of time privileged
code was active during jobs. However, this presents a problem since
there is no way to run privileged destructor or finalization type
operations after a job shell is complete. For instance, to use PAM for
session management, the IMP would need to call pam_open_session(3)
before invoking the job shell, then pam_close_session(3) once the
job shell exits. The current design does not allow for this since
the IMP is no longer resident when the job shell exits.

Change the behavior of the IMP so it lingers during the lifetime of
the job. The IMP now forks a copy of itself and execs the job shell
from the copy. The privileged IMP now waits for the job shell to exit,
and any privilieged finalization routines can now be run after the
job shell exits, but before the IMP exits.