imp: version: add feature line with +pam if PAM-enabled
Problem: There is not currently a way to tell if the IMP has been
compiled with PAM support.

Add a line in `flux-imp version` to indicate compiled in features.
For now this prints "+pam" if the IMP is PAM-enabled. Set up the
output to be easily extensible if future other features are added.