imp: exec: add PAM session management support
Problem: Currently, a user on the system instance of Flux
is system-limited by administrative locked memory and file
constraints. HPC applications and hardware, particularly MPI,
OpenMP, and Infiniband networking, require more locked memory than
the default limit.

In order for Flux to support this, the IMP needs to call into the PAM
stack while running as the root user, establish a session with user
credentials, and then transition back to the user. This is handled
in the exec subcommand.

Note this will not work without the --enable-pam configure flag and
a configured PAM session stack.

Fixes flux-framework/flux-pmix#40, #148