imp: kill: print pid info with signal 0
Problem: It will be useful, especially in test and debug, to get
information the flux-imp kill command gathers about a pid, such as
the cgroup path, PID owner and cgroup owner, but no such support
currently exists.

Add support for printing the PID info discovered by flux-imp kill
when the signal passed to the command is 0.