testsuite: test flux-imp kill of another flux-imp
Problem: No tests ensure that flux-imp kill functionality targets
the children of the IMP when the target is itself a flux-imp process
owned by UID 0.

Add a test that attempts to simulate this situation.

The test is unforunately a bit convoluted, since flux-imp kill
functionality requires that the target processes are running in a
cgroup which is owned by the requesting user (i.e. the user running
the tests), and for this particular case the IMP must also be running
as root (as would be the case in a system instance, but tricky to
produce reliably in any test situation).