imp: improve errno handling in pidinfo.c
Problem: Some of the code in pidinfo.c, which used to be internal to
imp/kill.c, needs to be updated to improve errno handling.

Save errno where necessary to ensure it is not clobbered.

Remove useless tests of snprintf(3) return code when formatted
string is guaranteed to fit in destination buffer.